AIFunctionApp/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-storage-blob
scikit-learn
pickle-mixin
spotipy
transformers
bs4
pydub
librosa
numpy
pandas
requests
torch==2.4.0
torchvision==0.19.0
-f https://download.pytorch.org/whl/cpu
